fix warning: Suspicious sequence of types castings: memsize -> 32-bit integer -> memsize.

+++ src/core/FileMasks.cpp
@@ -1285,7 +1285,7 @@ void TFileCustomCommand::ValidatePattern(const UnicodeString & Command,
 
   if (PatternCmd == L'&')
   {
-    Found[0] = static_cast<int>(Index);
+    Found[0] = Index;
   }
   else if ((PatternCmd != TEXT_TOKEN) && (PatternLen(Command, Index) == 1))
   {
